--- example/basic/1-simple-scala/.mill-version
@@ -1,1 +1,0 @@
-0.11.0-M8-36-d74aa6-DIRTY77746533
