+++ scalalib/src/mill/scalalib/Lib.scala
@@ -2,8 +2,9 @@ package mill
 package scalalib
 
 import coursier.util.Task
-import coursier.{Dependency, Repository, Resolution}
+import coursier.{Dependency, LocalRepositories, Repositories, Repository, Resolution}
 import mill.api.{Ctx, Loose, PathRef, Result}
+import mill.modules.Util
 import mill.scalalib.api.ZincWorkerUtil
 
 object Lib {
@@ -133,6 +134,48 @@ object Lib {
       path <- (if (os.isDir(root.path)) os.walk(root.path) else Seq(root.path))
       if os.isFile(path) && (extensions.exists(path.ext == _) && !isHiddenFile(path))
     } yield path
+  }
+
+  def resolveMillBuildDeps(
+      repos0: Seq[Repository],
+      ctx: Option[mill.api.Ctx.Log],
+      useSources: Boolean
+  ): Seq[os.Path] = {
+    Util.millProperty("MILL_BUILD_LIBRARIES") match {
+      case Some(found) => found.split(',').map(os.Path(_)).distinct.toList
+      case None =>
+        val repos = repos0 ++ Set(
+          LocalRepositories.ivy2Local,
+          Repositories.central
+        )
+        val millDeps = BuildInfo.millEmbeddedDeps.split(",").map(d => ivy"$d").map(dep =>
+          BoundDep(Lib.depToDependency(dep, BuildInfo.scalaVersion, ""), dep.force)
+        )
+        val Result.Success(res) = scalalib.Lib.resolveDependencies(
+          repositories = repos.toList,
+          deps = millDeps,
+          sources = useSources,
+          mapDependencies = None,
+          customizer = None,
+          coursierCacheCustomizer = None,
+          ctx = ctx
+        )
+
+        // Also trigger resolve sources, but don't use them (will happen implicitly by Idea)
+        if (!useSources) {
+          scalalib.Lib.resolveDependencies(
+            repositories = repos.toList,
+            deps = millDeps,
+            sources = true,
+            mapDependencies = None,
+            customizer = None,
+            coursierCacheCustomizer = None,
+            ctx = ctx
+          )
+        }
+
+        res.items.toList.map(_.path)
+    }
   }
 
 }
